+++ bmc-core/src/main/java/com/oracle/bmc/core/VirtualNetworkAsync.java
@@ -1728,6 +1728,23 @@ public interface VirtualNetworkAsync extends AutoCloseable {
                     handler);
 
     /**
+     * Get redundancy status of single DRG object on Oracle side.
+     *
+     *
+     * @param request The request object containing the details to send
+     * @param handler The request handler to invoke upon completion, may be null.
+     * @return A Future that can be used to get the response if no AsyncHandler was
+     *         provided. Note, if you provide an AsyncHandler and use the Future, some
+     *         types of responses (like java.io.InputStream) may not be able to be read in
+     *         both places as the underlying stream may only be consumed once.
+     */
+    java.util.concurrent.Future<GetDrgRedundancyStatusResponse> getDrgRedundancyStatus(
+            GetDrgRedundancyStatusRequest request,
+            com.oracle.bmc.responses.AsyncHandler<
+                            GetDrgRedundancyStatusRequest, GetDrgRedundancyStatusResponse>
+                    handler);
+
+    /**
      * Gets the specified provider service.
      * For more information, see [FastConnect Overview](https://docs.cloud.oracle.com/Content/Network/Concepts/fastconnect.htm).
      *