+++ bmc-core/src/main/java/com/oracle/bmc/core/VirtualNetworkClient.java
@@ -2764,6 +2764,35 @@ public class VirtualNetworkClient implements VirtualNetwork {
     }
 
     @Override
+    public GetDrgRedundancyStatusResponse getDrgRedundancyStatus(
+            GetDrgRedundancyStatusRequest request) {
+        LOG.trace("Called getDrgRedundancyStatus");
+        final GetDrgRedundancyStatusRequest interceptedRequest =
+                GetDrgRedundancyStatusConverter.interceptRequest(request);
+        com.oracle.bmc.http.internal.WrappedInvocationBuilder ib =
+                GetDrgRedundancyStatusConverter.fromRequest(client, interceptedRequest);
+        com.google.common.base.Function<javax.ws.rs.core.Response, GetDrgRedundancyStatusResponse>
+                transformer = GetDrgRedundancyStatusConverter.fromResponse();
+
+        final com.oracle.bmc.retrier.BmcGenericRetrier retrier =
+                com.oracle.bmc.retrier.Retriers.createPreferredRetrier(
+                        interceptedRequest.getRetryConfiguration(), retryConfiguration);
+        return retrier.execute(
+                interceptedRequest,
+                retryRequest -> {
+                    final com.oracle.bmc.retrier.TokenRefreshRetrier tokenRefreshRetrier =
+                            new com.oracle.bmc.retrier.TokenRefreshRetrier(
+                                    authenticationDetailsProvider);
+                    return tokenRefreshRetrier.execute(
+                            retryRequest,
+                            retriedRequest -> {
+                                javax.ws.rs.core.Response response = client.get(ib, retriedRequest);
+                                return transformer.apply(response);
+                            });
+                });
+    }
+
+    @Override
     public GetFastConnectProviderServiceResponse getFastConnectProviderService(
             GetFastConnectProviderServiceRequest request) {
         LOG.trace("Called getFastConnectProviderService");