+++ bmc-database/src/main/java/com/oracle/bmc/database/model/AutonomousDatabaseBackupSummary.java
@@ -127,6 +127,15 @@ public class AutonomousDatabaseBackupSummary {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("isRestorable")
+        private Boolean isRestorable;
+
+        public Builder isRestorable(Boolean isRestorable) {
+            this.isRestorable = isRestorable;
+            this.__explicitlySet__.add("isRestorable");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonIgnore
         private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
 
@@ -143,7 +152,8 @@ public class AutonomousDatabaseBackupSummary {
                             timeEnded,
                             lifecycleDetails,
                             databaseSizeInTBs,
-                            lifecycleState);
+                            lifecycleState,
+                            isRestorable);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
             return __instance__;
         }
@@ -161,7 +171,8 @@ public class AutonomousDatabaseBackupSummary {
                             .timeEnded(o.getTimeEnded())
                             .lifecycleDetails(o.getLifecycleDetails())
                             .databaseSizeInTBs(o.getDatabaseSizeInTBs())
-                            .lifecycleState(o.getLifecycleState());
+                            .lifecycleState(o.getLifecycleState())
+                            .isRestorable(o.getIsRestorable());
 
             copiedBuilder.__explicitlySet__.retainAll(o.__explicitlySet__);
             return copiedBuilder;
@@ -333,6 +344,12 @@ public class AutonomousDatabaseBackupSummary {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("lifecycleState")
     LifecycleState lifecycleState;
+
+    /**
+     * Indicates whether the backup can be used to restore the associated Autonomous Database.
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("isRestorable")
+    Boolean isRestorable;
 
     @com.fasterxml.jackson.annotation.JsonIgnore
     private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();