+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## 1.12.3 - 2020-01-21
+### Added
+- Support for getting DRG redundancy status in the Networking service
+- Support for cloning autonomous databases from backups in the Database service
+
 ## 1.12.2 - 2020-01-14
 ### Added
 - Support for a description field on route rules and security rules in the Networking service